Strategies added to orders
