+++ src/functions/investing/orders/options/update.js
@@ -40,7 +40,8 @@ export const main = handler(async (event) => {
 					openImpliedVolatility = :openImpliedVolatility,
 					closeImpliedVolatility = :closeImpliedVolatility,
 					profitLoss = :profitLoss,
-					signalList = :signalList,
+					strategyId = :strategyId,
+					strategyName = :strategyName,
           modifyDate = :modifyDate`,
           ExpressionAttributeValues: {
             ":ticker": order.ticker,
@@ -49,24 +50,25 @@ export const main = handler(async (event) => {
             ":orderSize": order.orderSize,
             ":openPrice": order.openPrice,
             ":closePrice": order.closePrice,
-            ":openUnderlyingPrice": order.openUnderlyingPrice || null,
-            ":closeUnderlyingPrice": order.closeUnderlyingPrice || null,
+            ":openUnderlyingPrice": order.openUnderlyingPrice,
+            ":closeUnderlyingPrice": order.closeUnderlyingPrice,
             ":strikePrice": order.strikePrice,
             ":contractType": order.contractType,
             ":tradeSide": order.tradeSide,
             ":contractExpirationDate": order.contractExpirationDate,
-            ":openDelta": order.openDelta || null,
-            ":closeDelta": order.closeDelta || null,
-            ":openGamma": order.openGamma || null,
-            ":closeGamma": order.closeGamma || null,
-            ":openVega": order.openVega || null,
-            ":closeVega": order.closeVega || null,
-            ":openTheta": order.openTheta || null,
-            ":closeTheta": order.closeTheta || null,
-            ":openImpliedVolatility": order.openImpliedVolatility || null,
-            ":closeImpliedVolatility": order.closeImpliedVolatility || null,
+            ":openDelta": order.openDelta,
+            ":closeDelta": order.closeDelta,
+            ":openGamma": order.openGamma,
+            ":closeGamma": order.closeGamma,
+            ":openVega": order.openVega,
+            ":closeVega": order.closeVega,
+            ":openTheta": order.openTheta,
+            ":closeTheta": order.closeTheta,
+            ":openImpliedVolatility": order.openImpliedVolatility,
+            ":closeImpliedVolatility": order.closeImpliedVolatility,
             ":profitLoss": order.profitLoss,
-            ":signalList": order.signalList,
+            ":strategyId": order.strategyId,
+            ":strategyName": order.strategyName,
             ":modifyDate": Date.now(),
           },
         },
