+++ src/functions/investing/orders/shares/create.js
@@ -28,7 +28,8 @@ export const main = handler(async (event) => {
             closePrice: order.closePrice,
             tradeSide: order.tradeSide,
             profitLoss: order.profitLoss,
-            signalList: order.signalList,
+            strategyId: order.strategyId,
+            strategyName: order.strategyName,
             createDate: Date.now(),
             modifyDate: Date.now(),
           },
