+++ src/functions/investing/orders/shares/update.js
@@ -24,7 +24,8 @@ export const main = handler(async (event) => {
             closePrice = :closePrice,
             tradeSide = :tradeSide,
             profitLoss = :profitLoss,
-						signalList = :signalList,
+						strategyId = :strategyId,
+					strategyName = :strategyName,
           	modifyDate = :modifyDate`,
           ExpressionAttributeValues: {
             ":ticker": order.ticker,
@@ -35,7 +36,8 @@ export const main = handler(async (event) => {
             ":closePrice": order.closePrice,
             ":tradeSide": order.tradeSide,
             ":profitLoss": order.profitLoss,
-            ":signalList": order.signalList,
+            ":strategyId": order.strategyId,
+            ":strategyName": order.strategyName,
             ":modifyDate": Date.now(),
           },
         },
