+++ src/functions/investing/orders/verticalSpreads/create.js
@@ -44,7 +44,8 @@ export const main = handler(async (event) => {
             openImpliedVolatility: order.openImpliedVolatility || null,
             closeImpliedVolatility: order.closeImpliedVolatility || null,
             profitLoss: order.profitLoss,
-            signalList: order.signalList,
+            strategyId: order.strategyId,
+            strategyName: order.strategyName,
             createDate: Date.now(),
             modifyDate: Date.now(),
           },
