+++ src/functions/investing/orders/verticalSpreads/update.js
@@ -40,7 +40,8 @@ export const main = handler(async (event) => {
             openImpliedVolatility = :openImpliedVolatility,
             closeImpliedVolatility = :closeImpliedVolatility,
             profitLoss = :profitLoss,
-            signalList = :signalList,
+            strategyId = :strategyId,
+						strategyName = :strategyName,
           	modifyDate = :modifyDate`,
           ExpressionAttributeValues: {
             ":ticker": order.ticker,
@@ -67,7 +68,8 @@ export const main = handler(async (event) => {
             ":openImpliedVolatility": order.openImpliedVolatility || null,
             ":closeImpliedVolatility": order.closeImpliedVolatility || null,
             ":profitLoss": order.profitLoss,
-            ":signalList": order.signalList,
+            ":strategyId": order.strategyId,
+            ":strategyName": order.strategyName,
             ":modifyDate": Date.now(),
           },
         },
